update expo sdk 40
