+++ package.json
@@ -45,7 +45,6 @@
     "@types/color": "^3.0.0",
     "@types/compare-versions": "^3.3.0",
     "@types/deep-equal": "^1.0.1",
-    "@types/expo": "^33.0.1",
     "@types/firebase": "^3.2.1",
     "@types/react": "~16.9.35",
     "@types/react-native": "~0.63.2",
