+++ src/components/organisms/Header/Header.tsx
@@ -1,16 +1,11 @@
 import React, { memo, useCallback } from 'react';
-import {
-  TouchableOpacity,
-  View,
-  Text,
-  StatusBar,
-  StyleSheet,
-} from 'react-native';
+import { TouchableOpacity, View, Text, StyleSheet } from 'react-native';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 import { useActionSheet } from '@expo/react-native-action-sheet';
 import theme from 'config/theme';
 import { IconImage } from 'components/atoms';
 import { getKindData } from 'lib/kind';
+import FocusAwareStatusBar from 'components/organisms/FocusAwareStatusBar';
 
 type Props = {
   date: string;
@@ -60,7 +55,10 @@ const Header: React.FC<Props> = (props) => {
         backgroundColor: props.color,
       }}
     >
-      <StatusBar backgroundColor={props.color} barStyle="dark-content" />
+      <FocusAwareStatusBar
+        backgroundColor={props.color}
+        barStyle="dark-content"
+      />
       <View style={styles.header}>
         <TouchableOpacity onPress={props.onClose}>
           <MaterialCommunityIcons
