+++ src/containers/Auth.tsx
@@ -1,7 +1,11 @@
 import * as GoogleSignIn from 'expo-google-sign-in';
 import * as Google from 'expo-google-app-auth';
-import { Platform } from 'react-native';
+import { Platform, Alert } from 'react-native';
 import AsyncStorage from '@react-native-community/async-storage';
+import uuidv4 from 'uuid/v4';
+import dayjs from 'dayjs';
+import advancedFormat from 'dayjs/plugin/advancedFormat';
+import 'dayjs/locale/ja';
 import * as AppleAuthentication from 'expo-apple-authentication';
 import React, {
   memo,
@@ -12,12 +16,26 @@ import React, {
   useContext,
   useEffect,
 } from 'react';
+import { View } from 'react-native';
+import EStyleSheet from 'react-native-extended-stylesheet';
+import Spinner from 'react-native-loading-spinner-overlay';
 import Constants from 'expo-constants';
+import { backup } from 'lib/backup';
 import * as Sentry from 'sentry-expo';
 import firebase from 'lib/system/firebase';
 import useIsFirstRender from 'hooks/useIsFirstRender';
 import { UID } from 'domain/user';
 import libAuth from 'lib/auth';
+import {
+  useSyncCalendarsMutation,
+  SyncCalendarsMutationVariables,
+  SyncCalendar,
+  SyncItemDetail,
+} from 'queries/api/index';
+import { post } from 'lib/fetch';
+import theme from 'config/theme';
+
+dayjs.extend(advancedFormat);
 
 const auth = new libAuth();
 
@@ -29,6 +47,7 @@ type Props = {};
 type State = {
   uid: UID;
   setup: boolean;
+  render: boolean;
   email: string;
 };
 
@@ -50,32 +69,129 @@ const Auth: FC<Props> = memo((props) => {
   const [state, setState] = useState<State>({
     email: '',
     uid: null,
+    render: true,
     setup: false,
   });
   const isFirstRender = useIsFirstRender();
-
-  const setSession = useCallback(async (refresh = false) => {
-    const idToken = await auth.setSession(refresh);
-
-    if (idToken) {
-      const user = firebase.auth().currentUser;
-      if (user) {
-        await AsyncStorage.setItem('email', user?.email || '');
-        await AsyncStorage.setItem('uid', user.uid);
-        setState((s) => ({
-          ...s,
-          email: user.email || '',
-          uid: user.uid || '',
-        }));
-      }
-    }
-
-    return idToken;
-  }, []);
+  const [syncCalendarsMutation] = useSyncCalendarsMutation({
+    async onCompleted() {
+      setRender();
+    },
+    onError(err) {
+      Alert.alert('バックアップに失敗しました', err.message);
+    },
+  });
 
   const getIdToken = useCallback(async () => {
     return await auth.getIdToken();
   }, []);
+
+  const setRender = useCallback(() => {
+    setState((s) => ({
+      ...s,
+      render: true,
+    }));
+  }, []);
+
+  const saveUser = useCallback(async () => {
+    const idToken = await getIdToken();
+
+    const response = await post('CreateUser', null, idToken || '');
+
+    if (response?.error) {
+      Alert.alert('ユーザーの保存に失敗しました。');
+    }
+
+    return Number(response?.status);
+  }, [getIdToken]);
+
+  const backupData = useCallback(async () => {
+    const httpStatus = await saveUser();
+
+    if (httpStatus === 201) {
+      // ユーザーを新規で作成した場合はデータをサーバーに送る
+      const { items, itemDetails, calendars } = await backup();
+
+      const syncCalendars = calendars.map((v) => {
+        const i = items.find((v1) => v1.id === v.itemId);
+        let ids: SyncItemDetail[] = [];
+
+        if (i) {
+          ids = itemDetails
+            .filter((v2) => v2.itemId === v.itemId)
+            .map((id) => ({
+              id: uuidv4(),
+              title: id.title || '',
+              kind: id?.kind || '',
+              place: id?.place || '',
+              url: id?.url || '',
+              memo: id?.memo || '',
+              priority: id?.priority || 1,
+            }));
+        }
+
+        const r: SyncCalendar = {
+          id: uuidv4(),
+          date: dayjs(v.date).format('YYYY-MM-DDT00:00:00'),
+          item: {
+            id: uuidv4(),
+            title: i?.title || '',
+            kind: i?.kind || '',
+            itemDetails: ids,
+          },
+        };
+
+        return r;
+      });
+
+      const variables: SyncCalendarsMutationVariables = {
+        calendars: {
+          calendars: syncCalendars,
+        },
+      };
+
+      syncCalendarsMutation({ variables });
+    } else if (httpStatus === 200) {
+      setRender();
+    } else {
+      Alert.alert('エラー', 'ユーザー登録/ログインに失敗しました', [
+        {
+          text: '閉じる',
+          onPress: () => {
+            setRender();
+          },
+        },
+      ]);
+    }
+  }, [saveUser, syncCalendarsMutation, setRender]);
+
+  const setSession = useCallback(
+    async (refresh = false) => {
+      const idToken = await auth.setSession(refresh);
+
+      if (idToken) {
+        const user = firebase.auth().currentUser;
+        if (user) {
+          setState((s) => ({
+            ...s,
+            render: false,
+          }));
+          await AsyncStorage.setItem('email', user?.email || '');
+          await AsyncStorage.setItem('uid', user.uid);
+          setState((s) => ({
+            ...s,
+            email: user.email || '',
+            uid: user.uid || '',
+          }));
+
+          await backupData();
+        }
+      }
+
+      return idToken;
+    },
+    [backupData]
+  );
 
   const loggedIn = useCallback(async () => {
     const idToken = await getIdToken();
@@ -225,16 +341,34 @@ const Auth: FC<Props> = memo((props) => {
   }, [isFirstRender, state, loggedIn]);
 
   const onLogout = useCallback(async () => {
+    setState((s) => ({
+      ...s,
+      render: false,
+    }));
+
     await logout();
     setState((s) => ({
       ...s,
       email: '',
       uid: null,
+      render: true,
     }));
   }, []);
 
   if (!state.setup) {
     return null;
+  }
+
+  if (!state.render) {
+    return (
+      <View style={styles.container}>
+        <Spinner
+          visible={true}
+          textContent="読込中"
+          textStyle={{ color: theme().color.white }}
+        />
+      </View>
+    );
   }
 
   return (
@@ -277,3 +411,11 @@ const nonceGen = (length: number) => {
   }
   return result;
 };
+
+const styles = EStyleSheet.create({
+  container: {
+    alignItems: 'center',
+    backgroundColor: '$background',
+    height: '100%',
+  },
+});
