+++ src/containers/__tests__/Auth.test.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { shallow, ShallowWrapper } from 'enzyme';
+import * as queries from 'queries/api/index';
 import Auth from '../Auth';
 
 describe('containers/Auth.tsx', () => {
@@ -10,6 +11,10 @@ describe('containers/Auth.tsx', () => {
   });
 
   beforeEach(() => {
+    jest
+      .spyOn(queries, 'useSyncCalendarsMutation')
+      .mockImplementation(() => [jest.fn()] as any);
+
     wrapper = shallow(<Auth {...propsData()} />);
   });
